companion-agent/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
aiofiles>=24.1.0
python-dotenv>=1.0.0
